requirements.txt
gradio
flask
google-api-python-client
google-auth-oauthlib
pandas
requests
python-dotenv
